Increase button padding
Using the assign button in the core app as a baseline, increase the
padding on the button to better match a button in the core app.

Technically the assign button in the core app is considered "wide" and
not standard, however the "wide" version of the button seems to be
commonly used in the core app so this makes the "wide" version standard
in Llama Goggles.